v.2.4.0
##New tab in the plugin parameters - Integrations
Added In the new tab "Integrations" there is a list of supported extensions that the plugin was able to detect on your site.
##Joomla Radical Form Support
If you use the professional Radical Form plugin on your site as a feedback form, then now this data will also get into your CRM. Documentation on the website of the Radical Form developer. In order for the feedback form data to get into the Bitrix24 CRM, name the fields of your feedback forms in the same way as the fields of Bitrix24. You can find an example of the form in the plugin parameters, the "Integrations" tab.
#Showing debugging information and cleaning the code
The mechanism for displaying debugging information has been completely redesigned. A small and not very good refactoring of the code was performed.